requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Async Support
# asyncio is built-in for Python 3.9+, no separate package needed
aiohttp==3.9.1

# Database
supabase>=2.11.0
postgrest>=0.13.2
asyncpg==0.29.0

# Redis
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==42.0.0

# Pydantic for validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.3.0

# Geospatial
geopy==2.4.1
python-geohash==0.8.5

# WebSockets
websockets>=15.0

# HTTP Client (version compatible with supabase 2.11+)
httpx>=0.24.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Monitoring & Logging
structlog==24.1.0
sentry-sdk[fastapi]==1.39.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
# httpx already included above for supabase compatibility

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Background Jobs (optional - add when needed)
# celery==5.3.4
# redis==5.0.1  # Already included above
# flower==2.0.1

# Production Server
gunicorn==21.2.0
